llmape/src/main/leaderboard/requirements.txt
jaydebeapi
psycopg2-binary
pandas
numpy
scikit-learn